+++ R/r_pcaSpectra.R
@@ -39,7 +39,8 @@
 #' pca <- r_pcaSpectra(metMUD1)
 #' plotScores(metMUD1, pca, main = "metMUD1 NMR Data",
 #'   pcs = c(1,2), ellipse = "cls", tol = 0.05)
-#' 
+#' plotScree(pca)
+#'
 #' @export r_pcaSpectra
 #'
 #'